ad_spdif: refine logic for DTS 44.1khz fix
There are some DTS variants with a samplerate of 96khz (e.g. DTS 24/96), somehow they are recognized wrongly as 48khz by the code, so it works also without this patch. Don´t rely on this "bug", make it correct. Now every samplerate above 44.1Khz is correctly treated as 48khz, and 44.1khz files are treated as 44.1khz for bitstreaming.